+++ libs/cardano-ledger-core/test/Main.hs
@@ -6,6 +6,7 @@ import qualified Test.Cardano.Ledger.BinarySpec as BinarySpec
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.JsonSpec as JsonSpec
 import qualified Test.Cardano.Ledger.PlutusSpec as PlutusSpec
+import qualified Test.Cardano.Ledger.State.StakePoolSpec as StakePoolSpec
 import qualified Test.Cardano.Ledger.ToolsSpec as ToolsSpec
 
 main :: IO ()
@@ -17,4 +18,5 @@ main =
       BinarySpec.spec
       JsonSpec.spec
       PlutusSpec.spec
+      StakePoolSpec.spec
       ToolsSpec.spec