+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/TreeDiff.hs
@@ -28,7 +28,6 @@ import Cardano.Ledger.Plutus.Data
 import Cardano.Ledger.Plutus.ExUnits
 import Cardano.Ledger.Plutus.Language
 import Cardano.Ledger.Plutus.TxInfo
-import Cardano.Ledger.PoolParams
 import Cardano.Ledger.State
 import Cardano.Ledger.TxIn
 import Cardano.Ledger.UMap
@@ -201,6 +200,8 @@ instance ToExpr CompactAddr
 instance ToExpr PoolMetadata
 
 instance ToExpr PoolParams
+
+instance ToExpr StakePoolState
 
 instance ToExpr StakePoolRelay