+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Conway/LedgerTypes/WellFormed.hs
@@ -71,7 +71,7 @@ dsX = do
   univ <- genWitUniv @era 100
   context <- genCertContext @era univ
   khppMap <- genFromSpec (witnessedKeyHashPoolParamMapSpec univ)
-  genFromSpec @(DState era) (conwayDStateSpec univ context (lit khppMap))
+  genFromSpec @(DState era) (conwayDStateSpec univ context (lit $ mkStakePoolState <$> khppMap))
 
 conwayDStateGen ::
   forall era. era ~ ConwayEra => Gen (DState era)